Update to use multi-stage builds
The switch to multi-stage builds is important in order to keep
image size low, while avoiding the pitfalls of previous
Dockerfile practices.

The relevant pitfalls of the old way include:

* image size would still remain large if dependencies are installed in
one RUN step and removed in another
* chaining all of those steps together in one long shell command is
hard to read and tedious to maintain
* moving the commands into a shell script (as done previously) could
obscure build details from an interested party by moving these steps
outside the Dockerfile

Although this makes the "quickstart" guide a little more tedious in
terms of template material, it should be a better solution overall.